Implementação dos testes de produtos

+++ src/main/java/br/com/techchallenge4/msprodutos/model/ProdutoEstoque.java
@@ -3,9 +3,11 @@ package br.com.techchallenge4.msprodutos.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 public class ProdutoEstoque {
     private Long id;
-    private Long qtdEstoque;
+    private BigInteger qtdEstoque;
 }
